Break details pages into components.
